pre-build proto structures #138
