--- src/tonic/proto.rs
@@ -1,1 +1,0 @@
-tonic::include_proto!("qdrant");
